Player data fixes
